Replace the overlap check of D struct initializers by an empty stub for C++ records.
VarDeclaration::offset is in bytes which isn't good for bitfields. We could do a real check with offset and sizes in bits but those overlap checks are already done anyway by Clang, so let's disable the check altogether for now.

Proper handling of bitfields in codegen is still needed (or should they be mapped to std.bitmanip? this would probably have unforeseen implications though).