Generate the PCH only after all the semantic phases are completed, in order to prevent redoing template and implicit member instantiations.
This provides an enormous speed boost to the semantic phases of C++ modules after the PCH generation.

Note that there is no easy way with Clang to save the entire AST into a PCH if a PCH is already loaded, so this is only possible once, new instantiations in subsequent compiler runs after the PCH generation won't be saved.

This also goes back to Clang's ASTUnit (modified to take a custom ASTConsumer) instead of Calypso's flavored stripped down version.