Make Clang emit callees with the always inline attribute.
Always inlined functions are an exception to df1140a474a0b7e3a60c634eeb8312355e060b01, in order to honor their inlining they have to be emitted in the caller module. Previously Calypso was tracking them down with the AST visitor, but doing it in EmitCall on Clang's side is faster and cleaner.