Change the codegen strategy towards something more DMD-like by considering only D modules to be truly "root", even if the __cpp modules get codegen'd.
If the __cpp modules were considered 'root', TemplateInstance::appendToModuleMember() would add template instances to the tempdecl's module members, not the instantiating module.

But not making them 'root' would skip codegen because the symbols would be thought to be external.

The __cpp modules are codegen'd only once and we want template instances from D to go into the instantiating D module, so a middle ground had to be found.

The emission of non-template functions needed a small fix as well.