Do not let DMD generate a ctor for C++ derived classes.
Is this the right thing to do?

If the ctor from a derived class isn't mapped by Calypso, this means that it's trivial and implicit, and the user-provided base ctor is a trivial one. So if this is the way to go, this should be turned into a virtual bool generateDerivedCtor() hook.

The problem this avoids is CppMangle not supporting every symbol mapped by Calypso.