Only semantic TemplateValueParameter.valType if valType isn't dependent on another parameter.
92965673e58c179cc6e8be3083f4e2e884fbeeed had to be reverted during the merge. Not semantic'ing valType it causes dummyArg() from leastAsSpecialized and in turn matchArg() to instantiate other template instantiations with sc.tinst pointing to a dummy "scope TemplateInstance", and thus tinst becomes a dangling pointer, which gets read later during TemplateInstance.needsCodegen().

But C++ templates may have value parameters with types dependent on a previous type parameter. Since dummyArg() is only called for D templates, assume that valType is non-dependent. Ideally the new nonDependentValType wouldn't be needed, but this sticks to the current behavior and ensures that D template value parameters never get a dependent type.