Move the Clang Driver initialization to cpp::LangPlugin::_init, add predefined version identifiers.
This is to determine which C++ standard library is selected early (which also receives a fix, it wasn't working before unless passed with -cc1), in order to define version identifiers before parsing :  "CppStdLib_libcxx" or "CppStdLib_libstdcxx".

However currently the clang::driver::Driver object has to be re-created after deleting the dummy clang::driver::Compilation for some unidentified reason, so this isn't ideal and needs to be investigated after the switch to newer Clang versions.