libc++ EH support, OS X build fix.
Compared to libstdc++, there are slight vtable and layout differences in the libc++ ABI:
    - std::type_info::can_catch has one less parameter
    - __cxa_exception/__cxa_dependent_exception have an extra referenceCount/primaryException field
    - caught pointer objects aren't expected to be deferenced i.e passed by value

Parts of libc++abi were copied, and libstdc++'s unwind-cxx.h file was also trimmed down. From now on the C++ header only depends on unwind.h, hence Calypso doesn't have to be built anymore against systemtap-sdt-dev and libstdc++.

Although it works on Linux, this doesn't fix issues #87 and #82 on MacOS yet, which seems to be due to ABI differences.

Thanks to @timotheecour for pointing out the existence of __shim_type_info.