Map nested records and lambda types with non-internal linkage before DeclReferencer traversal.
Those records are usually emitted by Clang's codegen when emitting the nesting function, but they may also be passed around as template arguments.

This can't be done by DeclReferencer because lambda classes may be used before and/or after they appear in the AST.

It's not possible to reach those nested symbols with lookups, so we rely on clang::Decl::dsym.