Dispatching templates part 1 (opUnary, opBinary and opOpAssign).
When searching for opBinary, instead of mapping every C++ function/template whose corresponding D identifier is opBinary, returns a "dispatching template" instead. The mapping of the actual symbols is done while instantiating this template, depending on the first explicit argument.

opCast support would be a nice addition, it remains to be implemented.