The type of a temporary for a function argument might be different from the type of the expression it gets initialized from, so bitcast it.
In functionParameters() arguments get cast to the parameter type, and this cast may be stripped later by optimize() (which simply alters the expression type instead). In the case of the in-place construction of a temporary through a function call with a sret argument the sret argument needs to have the correct type.